requirements.txt
coverage
linkchecker
pre-commit
pytest
apache-beam[gcp]>=2.53
sentence-transformers
google-cloud-bigquery
google-cloud-translate
pycld2
regex
yake
wordfreq
